Bug with Change for poll 420
When logging contacts it is not working as intended here.  I noticed that when I log a station in WSJT-X the Status wipes the callsign before the logged message goes through so the information is not there to save.  I'm not sure if this is the best way to address but it seems to work.  Not sure if it needs to be fixed a different way.