Update LogFormat.cpp
Made a change for LoTW if loading back date it will fill in missing information - only if the current record is blank.  I think this may be usefull for users that migrated from a different logger that didn't store as much information.  In running on mine just for 2024 it did add missing counties and IOTAs for instance.