Fixed #418 - Windows State/Size does not save in case of fullscreen (7)
Qt has an issue when performing saveGeometry in the case of a maximized window.
There are many workarounds, but none of them fully resolve the problem.
Therefore, this hack is used. If QLog detects a maximized window, it "changes" it as "normal" and
then resizes it to its original size. The user should not notice any difference in terms of size, but saveGeometry
works more reliably in this "normal" window mode.