Logbook: Filter highlighting
It often happens that users have a filter enabled
and when they enter a QSO outside of the filter range,
they report that they cannot see the QSO.
This needs to be highlighted more clearly.
The green color of the filter was not sufficient
so we will try using a red border instead.