UpdateDXSpotStatus
A quick example of what I would like to see. when I contact is logged update the DX Widget appropriately.  So you know that status is now worked.  Similar to what is currently done in the Alerts window.  I find like some others I keep the DXWidget and Alerts up for different things.