NewContact: RST LineEdit uses InputMask with Overwrite mode
New: If the default RST in Settings contains the '>' character,
it is treated as a pointer indicating the desired cursor focus position.

Change: The previous behavior of automatically jumping to the second position in the RST field has been removed.
Now, focus behavior depends on the presence of '>'; if absent, the cursor defaults to the first position.

This change was made in response to issue #695.