DXLab Commander Server
This is an initial attempt at a proof of concept.  We often see requests for ways to use other applications that need rig control, but if using rig control in QLog it is limited to the typical single connection of the serial port.  This adds a basic DXLab Commander Server to QLog so applications like wsjtx can be set to talk to port 52002 using the  DXLab Commander radio option.  I was curious if it would work and  this does seem to be a possible option to be considered.