Fixed #722 - Repeated log recording
This workaround is primarily due to JTDX. In general, with WSJTX, QLog receives
two messages for a single QSO. One is in the QByteArray format and the other in ADIF.
Both of these messages are received by QLog, and if they arrive within 0.5 seconds of each other,
they are merged and sent as a single QSO. If this does not happen, unfortunately, two QSOs are sent.

Under some mysterious circumstances, JTDX sometimes generates the second message after several seconds,
sometimes even up to 3 minutes. This workaround is intended to prevent duplicate entries
from being stored in the log.

In the future, however, this should be handled by merging the records.