Fixed #708 - CWConsole - Macro Buttons are Repeat Buttons
The Repeat Button works like a normal button, but it can
also learn an interval and press itself automatically. When
operator press the button (or use its shortcut) normally, it
performs its action once, just like any other button. If operator
then press the button again while holding the Shift key, the time
between the first press and the Shift-press becomes the repeat interval.
From that moment on, the button will trigger itself automatically at the learned
interval, and a colored bar inside the button will fill up to show the progress
until the next press.

To stop the automatic repeating, simply click the button again, press Esc, or
enter a callsign in NewContact.

If the repeating was stopped by any of the actions above, the learned interval
can be reactivated by pressing Shift + Shortcut again.