ui: Allow callbook lookups for WSJT-X contacts even without grid square
This change is to allow WSJT-X format broadcast packets to trigger a
callbook lookup even when the grid square is not specified. This is
useful for non FT8 content. For example, realtime contest logs being
sent from WriteLog to Qlog via the WSJT-X protocol.  It is safe to
call finalizeCallsignEdit() even without the grid square because the
CallbookManager will cache lookups and prevent redundant lookups
when WSJT-X sends multiple status type messages.