Support customizing the tile server for online map
In some countries with Internet censorship, the default OpenStreetMap tile
server may not be accessible. This commit adds an option to choose the preferred
tile server from a predefined list or customize them as needed.

By default, the online map retains the old behaviour which defaults to the user's
system language and it also provides an option to choose the langugage for OSM.

Signed-off-by: William Goodspeed <goodspeed@anche.no>